NetCDF writing optimization stage 1: - zero-copy expression values reading and netcdf writing - bye-bye netcdf-generator module and all its usages - no more expression values handling in DAO and Assay class - ArrayDesign objects now contain inverse map from DE accessions and names to their IDs - AtlasUnloader functionality moved to loader, no need for two separate modules here - AtlasLoader correctly cleans up old NetCDFs on experiment reloads
(Warning! This commit breaks tests)



git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@11518 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-web/src/main/java/ae3/dao/NetCDFReader.java
@@ -143,7 +143,7 @@ public class NetCDFReader {
          */
         experiment.setExpressionMatrix(arrayDesign, new ExpressionMatrix() {
             int lastDesignElement = -1;
-            double [] lastData = null;
+            float [] lastData = null;
             public double getExpression(int designElementId, int assayId) {
                 if(lastData != null && designElementId == lastDesignElement)
                     return lastData[assayId];
@@ -153,7 +153,7 @@ public class NetCDFReader {
                 originBDC[0] = designElementId;
                 shapeBDC[0] = 1;
                 try {
-                    lastData = (double[])varBDC.read(originBDC, shapeBDC).reduce().get1DJavaArray(double.class);
+                    lastData = (float[])varBDC.read(originBDC, shapeBDC).reduce().get1DJavaArray(float.class);
                 } catch(IOException e) {
                     throw new RuntimeException("Exception during matrix load", e);
                 } catch (InvalidRangeException e) {
